* system.h (getpagesize): Return 'int' instead of 'long'.
From-SVN: r88204

+++ gcc/system.h
@@ -303,8 +303,8 @@ extern char *getenv (const char *);
 extern int getopt (int, char * const *, const char *);
 #endif
 
-#if defined(HAVE_DECL_GETPAGESIZE) && !HAVE_DECL_GETPAGESIZE
-extern long getpagesize (void);
+#if defined (HAVE_DECL_GETPAGESIZE) && !HAVE_DECL_GETPAGESIZE
+extern int getpagesize (void);
 #endif
 
 #if defined (HAVE_DECL_GETWD) && !HAVE_DECL_GETWD
